Fix webpack-dev-server
Fixed the exhibition of the benchmark page. It wasn't even loading properly. Made minor changes to code to keep it more pleasant, like adding the current displayed table name on the screen.
