+++ benchmarks/styled-components/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import BigTable from './BigTable'
-// import Table from './ReactTable'
-// import Table from './SCTable'
-import Table from './XStyledTable'
+import BigTable from './components/BigTable'
+// import Table from './components/ReactTable'
+// import Table from './components/SCTable'
+import Table from './components/XStyledTable'
 
 export default () => <BigTable Table={Table} />
