+++ benchmarks/styled-components/src/index.js
@@ -2,7 +2,6 @@ import React from 'react'
 import { render } from 'react-dom'
 import App from './App'
 
-const root = document.createElement('div')
-document.body.appendChild(root)
+const root = document.querySelector('#root')
 
 render(<App />, root)
